+++ src/presenters/entity-page-projects.jsx
@@ -1,16 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {chunk, keyBy, partition} from 'lodash';
+import {partition} from 'lodash';
 
 import ProjectsList from './projects-list.jsx';
-import {CurrentUserConsumer, normalizeProjects} from './current-user.jsx';
-
 
 /* globals Set */
-
-function listToObject(list, val) {
-  return list.reduce((data, key) => ({[key]: val, ...data}), {});
-}
 
 const psst = "https://cdn.glitch.com/55f8497b-3334-43ca-851e-6c9780082244%2Fpsst.svg?1500486136908";
 
@@ -60,57 +54,4 @@ EntityPageProjects.propTypes = {
   addPin: PropTypes.func.isRequired,
   removePin: PropTypes.func.isRequired,
   projectOptions: PropTypes.object.isRequired,
-};
-
-export default class EntityPageProjectsLoader extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-  
-  ensureProjects(projects) {
-    const ids = projects.map(({id}) => id);
-    
-    const discardedProjects = Object.keys(this.state).filter(id => this.state[id] && !ids.includes(id));
-    if (discardedProjects.length) {
-      this.setState(listToObject(discardedProjects, undefined));
-    }
-    
-    const unloadedProjects = ids.filter(id => this.state[id] === undefined);
-    if (unloadedProjects.length) {
-      this.setState(listToObject(unloadedProjects, null));
-      chunk(unloadedProjects, 100).forEach(chunk => {
-        this.props.getProjects(chunk).then(projects => {
-          this.setState(keyBy(projects, ({id}) => id));
-        });
-      });
-    }
-  }
-  
-  componentDidMount() {
-    this.ensureProjects(this.props.projects);
-  }
-  
-  componentDidUpdate() {
-    this.ensureProjects(this.props.projects);
-  }
-  
-  render() {
-    const {projects, ...props} = this.props;
-    const loadedProjects = projects.map(project => this.state[project.id] || project);
-    return (
-      <CurrentUserConsumer>
-        {currentUser => (
-          <EntityPageProjects
-            projects={normalizeProjects(loadedProjects, currentUser)}
-            {...props}
-          />
-        )}
-      </CurrentUserConsumer>
-    );
-  }
-}
-EntityPageProjectsLoader.propTypes = {
-  projects: PropTypes.array.isRequired,
-  getProjects: PropTypes.func.isRequired,
 };
